schedule: add block/wake function
Background: channel is designed to be a data transfer pipe between
go routines. A scenario in gVisor is: block/wake a G (stands for a
task) by using channel, wherein there's no need for data transfer.
Channel is relatively too heavy for this case.
Channel manages Gs in a list, and has capacity. When the channel is
full, the sender will bock, and  the channel is clear, the receiver
is blocked. The receiver pushed to chan list, then schedule to run
next. The sender uses goready to wake Gs in the chan list.

A summary on our idea: introduce a new set of APIs for goroutine
wake/block.

Details:

We propose three new APIs:

GetG() -  used to get the address of goroutine.  By address, find
the goroutine G easier at go program.

WakeG() - can be used to wake one G, which can be in running/blocked
status.

BlockG() - can be used to block goroutine by itself.

How we use this in gVisor:

https://github.com/GuhuangLS/gvisor/commit/97e0e6c6cf861c46be6c6a2614d36d99db73a858

In futex()/epoll_wait(), we can modify it to use the new mechanism
for block and wake. Between sentry and go runtime, we maintain the
status of task Gs. Let's use futex as an example, add running status
at goruntime, NoWake,Waked,Blocked.

At sentry, one task/G can use BlockG() to block, like   <-chan. Other
tasks/Gs can use WakeG() to wake the task/G which is blocked by BlockG()
, like chan <-. Based on a basic prototype of Go and gVisor, we use the
program in google/gvisor#2033(comment) as the test program.

We can see 22% improvement by test case:
https://github.com/google/gvisor/issues/2033.

Signed-off-by: liushi <liushi.ls@antfin.com>